+++ lib/plug/conn/unfetched.ex
@@ -5,14 +5,18 @@ defmodule Plug.Conn.Unfetched do
   defstruct [:aspect]
 
   defimpl Access do
-    def get(%Plug.Conn.Unfetched{aspect: aspect}, key) do
+    def get(unfetched, key) do
+      raise_no_access(unfetched, key)
+    end
+
+    def get_and_update(unfetched, key, _value) do
+      raise_no_access(unfetched, key)
+    end
+
+    defp raise_no_access(%Plug.Conn.Unfetched{aspect: aspect}, key) do
       raise ArgumentError, message:
         "trying to access key #{inspect key} but they were not yet fetched. " <>
         "Please call Plug.Conn.fetch_#{aspect} before accessing it"
-    end
-
-    def access(unfetched, key) do
-      get(unfetched, key)
     end
   end
 end